build: 移除Debug配置并统一类型为float提升兼容性
- cvAutoTrack.vcxproj, TEST_cvAutoTrack_Cpp.vcxproj, cvAutoTrack.sln: 移除所有Debug|x64和Debug|Win32相关配置，精简项目配置
- cvAutoTrack.vcxproj, TEST_cvAutoTrack_Cpp.vcxproj: Release/RelNoOpz配置PlatformToolset由v145降为v143
- cvAutoTrack.vcxproj: nlohmann依赖路径由nlohmann\include改为nlohmann_json\include
- MatchType.h, Resources.h等: 结构体成员及常量由double统一为float，提升类型一致性
- tracking.cpp, Utils.cpp, KeypointsCache.cpp/h: 简化lowe_test、cleanAndComputePos_Old、KeypointGridLSH::build等接口，去除冗余参数
- AutoTrack.cpp, genshin.handle.cpp, genshin.minimap.cpp等: 相关计算和赋值同步float化
- window_scale.h, capture.include.h: 增加[[maybe_unused]]和(void)参数，消除未使用警告
- cvAutoTrack.cpp: LoadDependModuleFromPath增加TODO注释
- FileDownloaderAsync相关: 结构体成员初始化方式优化
- Release配置: 增加TreatAngleIncludeAsExternal和ExternalWarningLevel，优化外部头文件警告处理

BREAKING CHANGE: 移除了Debug配置，相关接口和结构体成员由double改为float，部分函数签名有变更，需同步适配调用方

+++ cvAutoTrack/src/AutoTrack.cpp
@@ -174,7 +174,7 @@ bool AutoTrack::SetWorldCenter(double x, double y)
 
 bool AutoTrack::SetWorldScale(double scale)
 {
-	genshin_avatar_position.target_map_world_scale = scale;
+	genshin_avatar_position.target_map_world_scale = static_cast<float>(scale);
 	return true;
 }
